challenges/web/Post It/service/post-it/app/requirements.txt
pyJWT[crypto]
flask